Update codebase with gridstudy
