+++ package.json
@@ -16,7 +16,7 @@
     "dependencies": {
         "@emotion/react": "^11.8.2",
         "@emotion/styled": "^11.8.1",
-        "@gridsuite/commons-ui": "^0.44.0",
+        "@gridsuite/commons-ui": "^0.46.0",
         "@hookform/resolvers": "^3.3.1",
         "@mui/icons-material": "^5.5.1",
         "@mui/lab": "^5.0.0-alpha.75",
