+++ src/services/study.ts
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import { backendFetchJson, Token } from '../utils/rest-api';
+import { backendFetchJson, Token } from '../utils/api-rest';
 
 const STUDY_URL = `${process.env.REACT_APP_API_GATEWAY}/study/v1`;
